Add support for lists with CIDR network specifications
Better parsing for different blocklists, one word at a time

increment version number and update read me

readme

+++ AutoBlockIPList.py
@@ -12,7 +12,7 @@ import time
 from functools import reduce
 
 
-VERSION = "1.0.1"
+VERSION = "1.1.0"
 
 
 def create_connection(db_file):
@@ -53,22 +53,60 @@ def ipv6_to_ipstd(ipv6):
 
 
 def process_ip(ip_list, expire):
-    processed = []
-    invalid = []
-    for i in ip_list:
-        try:
-            ip = ipaddress.ip_address(i)
+    processed_ips = set()
+    cidr_networks = set()
+    invalid = set()
+    total_potential_ips = 0
+    cidr_ip_count = 0
+
+    for line in ip_list:
+        for i in line.strip().split():
+            if not i or i.startswith("#") or "." not in i:
+                continue  # Skip empty lines and comments
+
+            try:
+                if '/' in i:
+                    network = ipaddress.ip_network(i)
+                    total_potential_ips += network.num_addresses
+                    if network.prefixlen >= 12:
+                        cidr_networks.add((str(network), expire))
+                        cidr_ip_count += network.num_addresses
+                    else:
+                        verbose(f"Network {i} is too large to be processed")
+                else:
+                    # Process as single IP
+                    ip = ipaddress.ip_address(i)
+                    if ip.version == 4:
+                        ipstd = ipv4_to_ipstd(i)
+                    elif ip.version == 6:
+                        ipstd = ipv6_to_ipstd(i)
+                    else:
+                        ipstd = ""
+                    processed_ips.add((i, ipstd, expire))
+                    total_potential_ips += 1
+
+            except ValueError:
+                if i != "":
+                    invalid.add(i)
+
+    return list(processed_ips), list(cidr_networks), list(invalid), cidr_ip_count
+
+def expand_cidr(cidr_str, expire):
+    expanded_ips = []
+    try:
+        network = ipaddress.ip_network(cidr_str)
+        for ip in network:
+            ip_str = str(ip)
             if ip.version == 4:
-                ipstd = ipv4_to_ipstd(i)
-            elif ip.version == 6:
-                ipstd = ipv6_to_ipstd(i)
+                ipstd = ipv4_to_ipstd(ip_str)
             else:
-                ipstd = ""
-            processed.append([i, ipstd, expire])
-        except ValueError:
-            if i != "":
-                invalid.append(i)
-    return processed, invalid
+                ipstd = ipv6_to_ipstd(ip_str)
+            expanded_ips.append((ip_str, ipstd, expire))
+
+    except ValueError as e:
+        verbose(f"Error expanding CIDR {cidr_str}: {e}")
+
+    return expanded_ips
 
 
 def url(link):
@@ -139,12 +177,15 @@ if __name__ == '__main__':
     # DSM 7: should be the same (TODO confirm path)
     db = "/etc/synoautoblock.db"
 
-    if not os.path.isfile(db):
+    db_present = os.path.isfile(db)
+    db_accessible = db_present and os.access(db, os.R_OK)
+    
+    if not args.dry_run and not db_present:
         raise FileNotFoundError(f"No such file or directory: '{db}'")
-    if not os.access(db, os.R_OK):
+    if not args.dry_run and not db_accessible:
         raise FileExistsError("Unable to read database. Run this script with sudo or root user.")
 
-    if args.backup_to is not None:
+    if args.backup_to is not None and db_present:
         filename = datetime.now().strftime("%Y%d%m_%H%M%S") + "_backup_synoautoblock.db"
         shutil.copy(db, os.path.join(args.backup_to, filename))
         verbose("Database successfully backup")
@@ -153,36 +194,53 @@ if __name__ == '__main__':
         args.expire_in_day = int(start_time) + args.expire_in_day * 60 * 60 * 24
 
     ips = get_ip_list(args.in_file, args.in_url)
-    ips_formatted, ips_invalid = process_ip(ips, args.expire_in_day)
+    ips_formatted, cidr_networks, ips_invalid, cidr_ip_count = process_ip(ips, args.expire_in_day)
+    verbose(f"IPs parsed from lists to be added: {len(ips_formatted)}")
+    verbose(f"Total CIDR networks: {len(cidr_networks)}")
+    verbose(f"Total IP in CIDR networks: {cidr_ip_count}")
+    verbose(f"Total IP invalid: {len(ips_invalid)}")
+    verbose(f"Total potential IP: {cidr_ip_count + len(ips_formatted)}")
 
-    verbose(f"Total IP fetch in lists: {len(ips_formatted)}")
+    if len(ips_formatted) + len(cidr_networks) > 0:
+        if db_accessible:
+            conn = create_connection(db)
+            c = conn.cursor()
 
-    if len(ips_formatted) > 0:
-        conn = create_connection(db)
-        c = conn.cursor()
-
-        if args.remove_expired and not args.dry_run:
+        if args.remove_expired and not args.dry_run and db_accessible:
             c.execute("DELETE FROM AutoBlockIP WHERE Deny = 1 AND ExpireTime > 0 AND ExpireTime < strftime('%s','now')")
             verbose("All expired entry was successfully removed")
 
-        if args.clear_db and not args.dry_run:
+        if args.clear_db and not args.dry_run and db_accessible:
             c.execute("DELETE FROM AutoBlockIP WHERE Deny = 1")
             verbose("All deny entry was successfully removed")
 
-        nb_ip = c.execute("SELECT COUNT(IP) FROM AutoBlockIP WHERE Deny = 1")
-        nb_ip_before = nb_ip.fetchone()[0]
+        if db_accessible:
+            nb_ip = c.execute("SELECT COUNT(IP) FROM AutoBlockIP WHERE Deny = 1")
+            nb_ip_before = nb_ip.fetchone()[0]
+            verbose(f"Total deny IP currently in your Synology DB: {nb_ip_before}")
+            
+        if len(ips_formatted) > 0:
+            if not args.dry_run and db_accessible:
+                c.executemany("REPLACE INTO AutoBlockIP (IP, IPStd, ExpireTime, Deny, RecordTime, Type, Meta) "
+                    "VALUES(?, ?, ?, 1, strftime('%s','now'), 0, NULL);", ips_formatted)
 
-        verbose(f"Total deny IP currently in your Synology DB: {nb_ip_before}")
-        if not args.dry_run:
-            c.executemany("REPLACE INTO AutoBlockIP (IP, IPStd, ExpireTime, Deny, RecordTime, Type, Meta) "
-                          "VALUES(?, ?, ?, 1, strftime('%s','now'), 0, NULL);", ips_formatted)
-        else:
-            verbose("Dry run -> nothing to do")
-        nb_ip = c.execute("SELECT COUNT(IP) FROM AutoBlockIP WHERE Deny = 1")
-        nb_ip_after = nb_ip.fetchone()[0]
-        conn.commit()
-        conn.close()
-        verbose(f"Total deny IP now in your Synology DB: {nb_ip_after} ({nb_ip_after - nb_ip_before} added)")
+        for cidr_str, expire in cidr_networks:
+            ips_formatted = expand_cidr(cidr_str, expire)
+            if len(ips_formatted) > 0 and not args.dry_run and db_accessible:
+                c.executemany("REPLACE INTO AutoBlockIP (IP, IPStd, ExpireTime, Deny, RecordTime, Type, Meta) "
+                    "VALUES(?, ?, ?, 1, strftime('%s','now'), 0, NULL);", ips_formatted)
+                
+        if db_accessible:
+            nb_ip = c.execute("SELECT COUNT(IP) FROM AutoBlockIP WHERE Deny = 1")
+            nb_ip_after = nb_ip.fetchone()[0]
+            conn.commit()
+            conn.close()
+            verbose(f"Total deny IP now in your Synology DB: {nb_ip_after} ({nb_ip_after - nb_ip_before} added)")
+        
+        if not db_accessible:
+            verbose("No database access. No changes were made to the database")
+        if args.dry_run:
+            verbose("Dry run mode. No changes were made to the database")
     else:
         verbose("No IP found in list")
